Fix file length violations after prettier formatting
Prettier reformatting expanded 7 files past the 150-line limit.
Split them further:
- session-manager.ts: extracted playback-operations.ts
- server/queue.js: extracted queue-helpers.js
- useSocketConnection.ts: extracted socketSetup.ts
- useSocketActions.ts: moved interfaces to handlerTypes.ts
- executors.ts: extracted fetchSongs.ts
- jellyfin-sdk/songs.ts: removed doc comments
