+++ src/hooks/search-actions/executors.ts
@@ -1,75 +1,15 @@
-import { Artist, Album, Playlist } from "@/types";
+import { fetchPlaylists, fetchArtists } from "@/services/searchFetchers";
 import {
-  fetchSongsByAlbum,
-  fetchSongsByArtist,
-  fetchPlaylists,
-  fetchSongsByPlaylist,
-  fetchArtists,
-} from "@/services/searchFetchers";
-import {
-  SongFetchContext,
   PlaylistFetchContext,
   ArtistFetchContext,
   deduplicateById,
 } from "./types";
 
-export async function executeFetchSongsByAlbum(
-  ctx: SongFetchContext,
-  album: Album,
-  page: number = 1,
-  append: boolean = false
-): Promise<void> {
-  page === 1 ? ctx.setIsLoading(true) : ctx.setIsLoadingMore(true);
-  if (page === 1) ctx.setSongResults([]);
-  ctx.setError(null);
-  try {
-    const result = await fetchSongsByAlbum(album.id, page);
-    if (append && page > 1) {
-      ctx.setSongResults(prev => deduplicateById([...prev, ...result.data]));
-    } else {
-      ctx.setSongResults(result.data);
-    }
-    ctx.setHasMoreResults(result.hasMore);
-    ctx.setCurrentPage(page);
-  } catch (err) {
-    const msg =
-      err instanceof Error ? err.message : "Failed to get songs by album";
-    ctx.setError(msg);
-    if (!append) ctx.setSongResults([]);
-  } finally {
-    ctx.setIsLoading(false);
-    ctx.setIsLoadingMore(false);
-  }
-}
-
-export async function executeFetchSongsByArtist(
-  ctx: SongFetchContext,
-  artist: Artist,
-  page: number = 1,
-  append: boolean = false
-): Promise<void> {
-  page === 1 ? ctx.setIsLoading(true) : ctx.setIsLoadingMore(true);
-  if (page === 1) ctx.setSongResults([]);
-  ctx.setError(null);
-  try {
-    const result = await fetchSongsByArtist(artist.id, page);
-    if (append && page > 1) {
-      ctx.setSongResults(prev => deduplicateById([...prev, ...result.data]));
-    } else {
-      ctx.setSongResults(result.data);
-    }
-    ctx.setHasMoreResults(result.hasMore);
-    ctx.setCurrentPage(page);
-  } catch (err) {
-    const msg =
-      err instanceof Error ? err.message : "Failed to get songs by artist";
-    ctx.setError(msg);
-    if (!append) ctx.setSongResults([]);
-  } finally {
-    ctx.setIsLoading(false);
-    ctx.setIsLoadingMore(false);
-  }
-}
+export {
+  executeFetchSongsByAlbum,
+  executeFetchSongsByArtist,
+  executeFetchSongsByPlaylist,
+} from "./fetchSongs";
 
 export async function executeFetchPlaylists(
   ctx: PlaylistFetchContext,
@@ -94,38 +34,10 @@ export async function executeFetchPlaylists(
     ctx.setHasMoreResults(result.hasMore);
     ctx.setCurrentPage(page);
   } catch (err) {
-    const msg = err instanceof Error ? err.message : "Failed to get playlists";
-    ctx.setError(msg);
+    ctx.setError(
+      err instanceof Error ? err.message : "Failed to get playlists"
+    );
     if (!append) ctx.setPlaylistResults([]);
-  } finally {
-    ctx.setIsLoading(false);
-    ctx.setIsLoadingMore(false);
-  }
-}
-
-export async function executeFetchSongsByPlaylist(
-  ctx: SongFetchContext,
-  playlist: Playlist,
-  page: number = 1,
-  append: boolean = false
-): Promise<void> {
-  page === 1 ? ctx.setIsLoading(true) : ctx.setIsLoadingMore(true);
-  if (page === 1) ctx.setSongResults([]);
-  ctx.setError(null);
-  try {
-    const result = await fetchSongsByPlaylist(playlist.id, page);
-    if (append && page > 1) {
-      ctx.setSongResults(prev => deduplicateById([...prev, ...result.data]));
-    } else {
-      ctx.setSongResults(result.data);
-    }
-    ctx.setHasMoreResults(result.hasMore);
-    ctx.setCurrentPage(page);
-  } catch (err) {
-    const msg =
-      err instanceof Error ? err.message : "Failed to get songs by playlist";
-    ctx.setError(msg);
-    if (!append) ctx.setSongResults([]);
   } finally {
     ctx.setIsLoading(false);
     ctx.setIsLoadingMore(false);
@@ -150,8 +62,7 @@ export async function executeFetchArtists(
     ctx.setHasMoreResults(result.hasMore);
     ctx.setCurrentPage(page);
   } catch (err) {
-    const msg = err instanceof Error ? err.message : "Failed to load artists";
-    ctx.setError(msg);
+    ctx.setError(err instanceof Error ? err.message : "Failed to load artists");
     if (!append) ctx.setArtistResults([]);
   } finally {
     ctx.setIsLoading(false);
